// console/packages/starwhale-ui/src/base/select/types.ts
export type SelectOverrides = {
    Root?: Override
    ControlContainer?: Override
    Placeholder?: Override
    ValueContainer?: Override
    SingleValue?: Override
    MultiValue?: Override
    Tag?: Override
    InputContainer?: Override
    Input?: Override
    IconsContainer?: Override
    SelectArrow?: Override
    ClearIcon?: Override
    LoadingIndicator?: Override
    SearchIconContainer?: Override
    SearchIcon?: Override
    Popover?: Override
    DropdownContainer?: Override
    Dropdown?: Override
    DropdownOption?: Override
    DropdownListItem?: Override
    OptionContent?: Override
    StatefulMenu?: Override
    StyledClearIcon?: Override
}
// ...
export type SelectProps = {
// ...
    'overrides'?: SelectOverrides
// ...
}